Rendering boxes

+++ core/src/com/ridwanstandingby/ntris/render/fonts/FontGenerator.kt
@@ -9,13 +9,13 @@ class FontGenerator(private val dimensions: Dimensions) {
     fun generate(): Fonts {
         val generator = FreeTypeFontGenerator(Gdx.files.internal(FontParameters.FONT_PATH))
         val parameter = FreeTypeFontGenerator.FreeTypeFontParameter()
-        parameter.size = dimensions.rescale(FontParameters.FONT_SMALL_SIZE)
+        parameter.size = dimensions.rescale(FontParameters.FONT_SMALL_SIZE).toInt()
         val fontSmall = generator.generateFont(parameter)
         parameter.borderWidth = FontParameters.FONT_BORDER_WIDTH
         parameter.borderColor = FontParameters.FONT_BORDER_COLOUR
         val fontSmallBordered = generator.generateFont(parameter)
         parameter.borderWidth = 0f
-        parameter.size = dimensions.rescale(FontParameters.FONT_LARGE_SIZE)
+        parameter.size = dimensions.rescale(FontParameters.FONT_LARGE_SIZE).toInt()
         val fontLarge = generator.generateFont(parameter)
         generator.dispose()
         return Fonts(fontSmall, fontLarge, fontSmallBordered)
